requirements.txt
streamlit
psycopg2-binary
openai
python-dotenv
pandas
